Formatting Changes, refactor
