+++ src/App.tsx
@@ -11,9 +11,10 @@
 import React, { Component } from "react";
 import { Platform, StyleSheet, Text, View } from "react-native";
 
-import Connect from "./src/components/connect/Connect";
+import Colors from "./colors";
+import Connect from "./components/connect/Connect";
 
-interface Props { }
+interface Props {}
 export default class App extends Component<Props> {
   render() {
     return (
@@ -29,7 +30,7 @@ const styles = StyleSheet.create({
     flex: 1,
     justifyContent: "center",
     alignItems: "center",
-    backgroundColor: "#F5FCFF"
+    backgroundColor: Colors.darkGrey
   },
   welcome: {
     fontSize: 20,
