+++ src/colors.tsx
@@ -1,0 +1,6 @@
+const Colors = {
+  darkGrey: "#0F0F14",
+  offWhite: "#F5FCFF"
+};
+
+export default Colors;
