+++ src/components/connect/Connect.tsx
@@ -1,6 +1,8 @@
 import React, { Component } from "react";
 import { Text, View, TextInput, Button, StyleSheet } from "react-native";
 
+import Colors from "../../colors";
+
 export default class Connect extends Component {
   state = {
     host: String,
@@ -17,15 +19,17 @@ export default class Connect extends Component {
 
   render() {
     return (
-      <View>
+      <View style={styles.body}>
         <Text style={styles.heading}>Kite</Text>
-        <Text>Enter Server Hostname or IP</Text>
+        <Text style={styles.info}>Enter Server Hostname or IP</Text>
         <TextInput
+          style={styles.input}
           placeholder="Hostname"
           onChangeText={host => this.updateHost(host)}
           keyboardType={"url"}
         />
         <TextInput
+          style={styles.input}
           placeholder="Port Number"
           onChangeText={port => this.updatePort(port)}
           keyboardType={"number-pad"}
@@ -37,15 +41,26 @@ export default class Connect extends Component {
 }
 
 const styles = StyleSheet.create({
+  body: {
+    backgroundColor: Colors.darkGrey,
+    width: "60%"
+  },
   heading: {
-    fontSize: 20,
+    fontSize: 30,
     fontWeight: "bold",
-    paddingBottom: 20
+    paddingBottom: 20,
+    color: Colors.offWhite
   },
   input: {
-    borderColor: "#2d2d2d",
-    borderWidth: 0.5,
     borderRadius: 4,
-    paddingVertical: 5
+    padding: 5,
+    marginBottom: 10,
+    backgroundColor: Colors.offWhite,
+    fontSize: 14
+  },
+  info: {
+    fontSize: 12,
+    paddingBottom: 10,
+    color: Colors.offWhite
   }
 });
